fix: Исправления П1: * Значения для создания целеуказаний указываются в градусах. * Общий список целеуказаний * Отображение верных чисел азимута и угла места
П2:
* Исправление мапинга

+++ AppOne/src/common.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct TargetDesignation {
-    short azimut;
-    short elev;
+    double azimut;
+    double elev;
 };
